+++ src/config.py
@@ -4,8 +4,6 @@ CW_BASE_URL = "https://crowdworks.jp"
 
 CW_JOB_DETAIL_URL = f"{CW_BASE_URL}/public/jobs"
 
-CW_JOB_LIST_URL = f"{CW_BASE_URL}/public/jobs/group/ai_bpo"
-
 OUTPUT_DIR = "output"
 
 DEBUG_DIR = "debug"